Life system implemented, lose state implemented
There is now a life system for the game. When a rock hit the sloth and
his branch it will subtract a life from the player. When the player
loses all their lives the gameover screen will display. Currently The
only way to restart the game is to refresh the browser.